# [fix] Faster data parsing