--- a/lib/client/multiplex.js
+++ b/lib/client/multiplex.js
@@ -45,10 +45,10 @@ module.exports = function multiplex(Spark) {
     var mp = this;
     this.conn.on('data', function ondata(data) {
       if (isArray(data)) {
-        var type = data.shift()
-          , id = data.shift()
-          , name = data.shift()
-          , payload = data.shift()
+        var type = data[0]
+          , id = data[1]
+          , name = data[2]
+          , payload = data.length === 4 ? data[3] : undefined
           , channel = mp.channels[name][id];
 
         if (!channel) return false;