--- a/lib/server/multiplex.js
+++ b/lib/server/multiplex.js
@@ -96,10 +96,10 @@ Multiplex.prototype.onconnection = function onconnection(conn) {
     if (!isArray(data)) return false;
 
     // Parse data to get required fields.
-    var type = data.shift()
-      , id = data.shift()
-      , name = data.shift()
-      , payload = data.shift()
+    var type = data[0]
+      , id = data[1]
+      , name = data[2]
+      , payload = data.length === 4 ? data[3] : undefined
       , channel = mp.channels[name];
 
     if (!channel) {