`shift()` modifies the incoming array which should be considered a bug (subsequent data listeners will get an empty array depending on where they are in the iteration order). It's also slow.

Testing on regular & out of bounds access (e.g. no `payload`, so data[3] doesn't exist), 10M iterations:
```
shift: 1618.473ms
index: 68.041ms
shift oob: 1607.713ms
index oob: 104.179ms
```